+++ app/components/base/table/index.ts
@@ -1,4 +1,5 @@
 export * from "./table.component";
+export * from "./table-head.component";
 export * from "./table-row.component";
 export * from "./table-cell.component";
 export * from "./table-column.component";
