+++ app/components/base/table/table.component.ts
@@ -1,60 +1,41 @@
 import {
-    AfterViewInit, ChangeDetectorRef, Component, ContentChild, ContentChildren, Optional, QueryList,
+    ChangeDetectorRef, Component, ContentChild, ContentChildren, Input, Optional, QueryList,
 } from "@angular/core";
 import { Router } from "@angular/router";
 
 import { FocusSectionComponent } from "app/components/base/focus-section";
 import { log } from "app/utils";
-import { AbstractListBase } from "../abstract-list";
+import { AbstractListBase, AbstractListBaseConfig, abstractListDefaultConfig } from "../abstract-list";
 import { SortDirection, TableColumnComponent } from "./table-column.component";
+import { TableHeadComponent } from "./table-head.component";
 import { TableRowComponent } from "./table-row.component";
-
 import "./table.scss";
 
-@Component({
-    selector: "bl-thead",
-    template: `<tr><th style="width: 30px"></th><ng-content></ng-content></tr>`,
-})
-export class TableHeadComponent implements AfterViewInit {
-    @ContentChildren(TableColumnComponent)
-    public items: QueryList<TableColumnComponent>;
-
-    private _columnIndexMap: StringMap<number>;
-
-    public ngAfterViewInit() {
-        this.items.changes.subscribe(() => {
-            this._updateColumnIndexMap();
-        });
-        this._updateColumnIndexMap();
-    }
-
-    public getColumnIndex(column: TableColumnComponent) {
-        if (!(column.id in this._columnIndexMap)) {
-            return -1;
-        }
-        return this._columnIndexMap[column.id];
-    }
-
-    private _updateColumnIndexMap() {
-        const map = {};
-        this.items.forEach((column, index) => {
-            map[column.id] = index;
-        });
-        this._columnIndexMap = map;
-    }
+export interface TableConfig extends AbstractListBaseConfig {
+    /**
+     * If it should show the checkbox for selected rows
+     * @default false
+     */
+    showCheckbox?: boolean;
 }
 
+export const tableDefaultConfig = {
+    ...abstractListDefaultConfig,
+    showCheckbox: false,
+};
 @Component({
     selector: "bl-table",
     templateUrl: "table.html",
 })
 export class TableComponent extends AbstractListBase {
-    @ContentChild(TableHeadComponent)
-    public head: TableHeadComponent;
+    @Input() public set config(config: TableConfig) {
+        this._config = { ...tableDefaultConfig, ...config };
+    }
 
-    @ContentChildren(TableRowComponent)
-    public items: QueryList<TableRowComponent>;
+    @ContentChild(TableHeadComponent) public head: TableHeadComponent;
+    @ContentChildren(TableRowComponent) public items: QueryList<TableRowComponent>;
 
+    protected _config: TableConfig;
     private _sortingColumn: TableColumnComponent;
 
     /**
